requirements.txt
tensorflow>=2.13
flask
pillow
numpy
scipy